--- src/Metrics.Web/Pages/Manage/KeyMetrics/Index.cshtml
@@ -1,6 +1,0 @@
-﻿@page
-@model Metrics.Web.Pages.Manage.KeyMetrics.IndexModel
-@{
-}
-
-<h3>Key Metrics List</h3>
