--- src/Metrics.Web/Pages/Manage/KeyMetrics/Index.cshtml.cs
@@ -1,12 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace Metrics.Web.Pages.Manage.KeyMetrics
-{
-    public class IndexModel : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
