+++ src/Metrics.Web/Pages/Manage/Submissions/KeyMetrics/Index.cshtml
@@ -1,0 +1,6 @@
+﻿@page
+@model Metrics.Web.Pages.Manage.Submissions.KeyMetrics.IndexModel
+@{
+}
+
+<h3>Key Metrics List</h3>
