+++ src/Metrics.Web/Pages/Manage/Submissions/KeyMetrics/Index.cshtml.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace Metrics.Web.Pages.Manage.Submissions.KeyMetrics;
+
+public class IndexModel : PageModel
+{
+    public void OnGet()
+    {
+    }
+}
