+++ src/Metrics.Web/Pages/Manage/Users/Register/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Metrics.Web.Pages.Manage.Users.RegisterModel
+@model Metrics.Web.Pages.Manage.Users.Register.IndexModel
 @{
     ViewData["Title"] = "Register new Account";
 }
