+++ src/Metrics.Web/Pages/Manage/Users/Register/Success.cshtml
@@ -1,5 +1,5 @@
 @page
-@model Metrics.Web.Pages.Manage.Users.SuccessModel
+@model Metrics.Web.Pages.Manage.Users.Register.SuccessModel
 @{
     ViewData["Title"]= "Registration Successful";
 }
