+++ src/Metrics.Web/Pages/Manage/Users/Roles/Create.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Metrics.Web.Pages.Manage.Roles.CreateModel
+@model Metrics.Web.Pages.Manage.Users.Roles.CreateModel
 @{
 }
