+++ src/Metrics.Web/Pages/Manage/Users/Roles/Create.cshtml.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Metrics.Web.Pages.Manage.Roles;
+namespace Metrics.Web.Pages.Manage.Users.Roles;
 
 public class CreateModel : PageModel
 {
