+++ src/Metrics.Web/Pages/Manage/Users/Roles/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Metrics.Web.Pages.Manage.Roles.IndexModel
+@model Metrics.Web.Pages.Manage.Users.Roles.IndexModel
 @{
      Layout = "/Pages/Shared/_Layout.cshtml";
 }
